Create fibonacci
Testase 1: For the given string, C={a,b,c} and f(a)=4,f(b)=1,f(c)=3. f(a)=f(c)+f(b) so the permutation (b,c,a) satisfies the requirement.
Testcase 2: Here too C={a,b,c} but no permutation satisfies the requirement of a dynamic string.
Testcase 3: Here C={m,n,o,p} and (m,n,o,p) is a permutation that makes it a dynamic string.